+++ src/browser/index.ts
@@ -2,3 +2,4 @@ export * from '../common';
 export * from './getElementBoundaries';
 export * from './getScrollbarWidth';
 export * from './closestPierce'
+export * from './disableEnableElementScroll'