+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 - `closestPierce` in browser utils
 - `getScrollbarWidth` options in browser utils
+- `disableElementScroll` in browser utils
+- `enableElementScroll` in browser utils
 
 ## [2.5.1]